LayoutTests/webgpu/whlsl/use-undefined-variable-expected.txt

CONSOLE MESSAGE: GPUDevice.createComputePipeline(): WHLSL compile error: 6:9: Cannot find the variable declaration.
PASS
PASS successfullyParsed is true

TEST COMPLETE
